separate current project build cache by interface modules https://online.sbis.ru/opendoc.html?guid=09d18934-8bd1-44a3-a890-5df5bfc045dd
