+++ gulp/builder/classes/modules-cache.js
@@ -6,6 +6,8 @@
 'use strict';
 
 const helpers = require('../../../lib/helpers');
+const path = require('path');
+const fs = require('fs-extra');
 
 function setDefaultStore() {
    return {
@@ -196,4 +198,48 @@ class ModuleCache {
    }
 }
 
-module.exports = ModuleCache;
+/**
+ * Read cache from disk if it exists
+ * @param moduleCachePath - path to current cache
+ * @returns {Promise<null>}
+ */
+async function getLastModuleCache(moduleCachePath) {
+   if (await fs.pathExists(moduleCachePath)) {
+      const result = await fs.readJson(moduleCachePath);
+      return result;
+   }
+   return null;
+}
+
+/**
+ * Task for getting saved module cache from disk if needed
+ * @param taskParameters - whole parameters list of current project build
+ * @param moduleInfo - main info about current module
+ * @returns {downloadModuleCache}
+ */
+function generateDownloadModuleCache(taskParameters, moduleInfo) {
+   moduleInfo.cachePath = path.join(taskParameters.config.cachePath, 'modules-cache', `${moduleInfo.name}.json`);
+   const patchBuild = taskParameters.config.modulesForPatch && taskParameters.config.modulesForPatch.length > 0;
+   return async function downloadModuleCache() {
+      const lastCache = await getLastModuleCache(moduleInfo.cachePath);
+      moduleInfo.cache = new ModuleCache(lastCache);
+      if (patchBuild && lastCache) {
+         /**
+          * in patch for module with rebuild configuration we need to get new themes meta from the
+          * last cache value
+          * In patch build new themes interface modules may not be participating in patch build,
+          * so we can lose meta for this modules.
+          */
+         if (moduleInfo.rebuild) {
+            const currentStoreModulesCache = taskParameters.cache.currentStore;
+            const lastStoreModulesCache = taskParameters.cache.lastStore;
+            currentStoreModulesCache.newThemesModules = lastStoreModulesCache.newThemesModules;
+         } else {
+            // in patch for modules without rebuild configuration store cache "as is"
+            moduleInfo.cache.currentStore = moduleInfo.cache.lastStore;
+         }
+      }
+   };
+}
+
+module.exports = generateDownloadModuleCache;
